% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcs.R
\name{arc_height_rules}
\alias{arc_height_rules}
\title{Arc height rules}
\usage{
arc_height_rules()
}
\value{
A character vector of rule names.
}
\description{
How a junction's count is turned into the height of its arc.
}
\examples{
arc_height_rules()
}
