% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-tables.R
\name{as_sashimi_data}
\alias{as_sashimi_data}
\title{Coerce an object to sashimi data}
\usage{
as_sashimi_data(x, ...)
}
\arguments{
\item{x}{A \code{sashimi_data} object, a named list of slot data frames, or a
path to a directory of tables.}

\item{...}{Passed to \code{\link[=sashimi_from_tables]{sashimi_from_tables()}}.}
}
\value{
A \code{sashimi_data} object.
}
\description{
Used by the plotting functions so they accept a \code{sashimi_data} object, a
list of slot tables, or a directory path interchangeably.
}
\examples{
as_sashimi_data(list(loci = data.frame(locus_id = "a", gene_name = "A",
  chrom = "chr1", strand = "+", win_lo = 1, win_hi = 10)))
}
