% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-rmats.R
\name{event_types}
\alias{event_types}
\title{Alternative splicing event types}
\usage{
event_types()
}
\value{
A character vector: \code{SE}, \code{A5SS}, \code{A3SS}, \code{MXE}, \code{RI}.
}
\description{
The five event classes rMATS reports.
}
\examples{
event_types()
}
