% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_coord}
\alias{format_coord}
\title{Format a genomic coordinate}
\usage{
format_coord(v, big_mark = ",")
}
\arguments{
\item{v}{Numeric vector of coordinates.}

\item{big_mark}{Thousands separator.}
}
\value{
A character vector.
}
\description{
Format a genomic coordinate
}
\examples{
format_coord(27035000)
}
