% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-contract.R
\name{loci}
\alias{loci}
\title{Locus identifiers held by a sashimi data object}
\usage{
loci(x)
}
\arguments{
\item{x}{A \code{sashimi_data} object.}
}
\value{
A character vector of \code{locus_id} values.
}
\description{
Locus identifiers held by a sashimi data object
}
\examples{
loci(sashimi_data())
}
