% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_methods}
\alias{normalize_methods}
\title{Coverage normalisation methods}
\usage{
normalize_methods()
}
\value{
A character vector of method names accepted by \code{\link[=normalize_tracks]{normalize_tracks()}}
and the \code{normalize} argument of \code{\link[=plot_sashimi]{plot_sashimi()}}.
}
\description{
Coverage normalisation methods
}
\examples{
normalize_methods()
}
