% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{omakase_palettes}
\alias{omakase_palettes}
\title{Palette names available}
\usage{
omakase_palettes()
}
\value{
A character vector of palette names.
}
\description{
Palette names available
}
\examples{
omakase_palettes()
}
