% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_colour_omakase}
\alias{scale_colour_omakase}
\alias{scale_color_omakase}
\alias{scale_fill_omakase}
\title{Discrete colour and fill scales using the omakase palettes}
\usage{
scale_colour_omakase(palette = "omakase", ...)

scale_color_omakase(palette = "omakase", ...)

scale_fill_omakase(palette = "omakase", ...)
}
\arguments{
\item{palette}{Palette name, see \code{\link[=omakase_palettes]{omakase_palettes()}}.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete colour and fill scales using the omakase palettes
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point() +
  scale_colour_omakase()
}
