% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{shrink_methods}
\alias{shrink_methods}
\title{Intron compression methods}
\usage{
shrink_methods()
}
\value{
A character vector of method names accepted by \code{\link[=intron_map]{intron_map()}} and by
the \code{shrink_method} argument of \code{\link[=plot_sashimi]{plot_sashimi()}}.
}
\description{
Intron compression methods
}
\examples{
shrink_methods()
}
