% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-bam.R
\name{strand_modes}
\alias{strand_modes}
\title{Strand specificity settings}
\usage{
strand_modes()
}
\value{
A character vector of setting names.
}
\description{
How the strand of a read is inferred from the flag, for stranded libraries.
}
\examples{
strand_modes()
}
