% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tracks.R
\name{track_axis}
\alias{track_axis}
\title{Build a coordinate axis track}
\usage{
track_axis(
  unit = c("auto", "bp", "Kb", "Mb"),
  n = 6,
  show_chrom = TRUE,
  height = 0.42
)
}
\arguments{
\item{unit}{\code{"auto"}, \code{"bp"}, \code{"Kb"} or \code{"Mb"}.}

\item{n}{Target number of ticks.}

\item{show_chrom}{Print the contig name under the axis.}

\item{height}{Relative height of this track.}
}
\value{
An \code{omakase_track} object.
}
\description{
Build a coordinate axis track
}
\examples{
track_axis()
}
