% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tracks.R
\name{track_spacer}
\alias{track_spacer}
\title{Build a spacer track}
\usage{
track_spacer(height = 0.15)
}
\arguments{
\item{height}{Relative height of the gap.}
}
\value{
An \code{omakase_track} object.
}
\description{
Build a spacer track
}
\examples{
track_spacer(0.1)
}
